backend/requirements.txt
asgiref==3.5.2
attrs==24.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.4.1
packaging==24.1
pytz==2022.6
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
